Split interpret.c into two files
Reduce the monstrous size of interpret.c by splitting
off the CTFE-specific expression types. No changes to code.
